BackEnd/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==5.1
django-cors-headers==4.4.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
dnspython==2.6.1
email_validator==2.2.0
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
idna==3.7
oauthlib==3.2.2
packaging==24.1
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
